feat: :sparkles: Flightconnection and city persistence added
MySQLFlightConenction created and City entity connection created.

@trainingLeader

+++ src/main/java/com/vuelosjanbi/city/infrastructure/adapters/in/CityConsoleAdapter.java
@@ -59,7 +59,7 @@ public class CityConsoleAdapter {
                         System.out.printf("ID: %d =, Name: %s\n", City.getId(), City.getCityName());
                     });
 
-                    int deleteCityId = scanner.nextInt();
+                    Long deleteCityId = scanner.nextLong();
                     scanner.nextLine();
 
                     CityService.deleteCity(deleteCityId);
